Convert the repo to a mono-repo
