+++ packages/react/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "@vybenetwork/react",
+  "version": "0.1.0",
+  "description": "React integration of the Vybe SDK",
+  "main": "src/index.ts",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/vybenetwork/vybe-js-sdk.git"
+  },
+  "keywords": [
+    "VybeNetwork",
+    "React",
+    "SDK"
+  ],
+  "author": "Vybe Network",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/vybenetwork/vybe-js-sdk/issues"
+  },
+  "homepage": "https://github.com/vybenetwork/vybe-js-sdk#readme"
+}
